Create methods and their contents/functionalities
Create functionalities for the ManageTransaction method; to build request using TerminalManageBuilder.
Create BuildManageTransaction and BuildRequestMessage to parse the request and send it to Ingenico based on what request procedure they expect.
Create IsObjectNullOrEmpty to identify whether the object is null or empty, ValidateExtendedData to validate the extended data used by the user, ReportRequest to process the request in /XMLReport transactions.